api/src/requirements.txt
Django
djangorestframework
django-rest-auth
django-allauth
django-filter
Pillow
bcrypt
django-cors-headers
channels
channels-redis
sendgrid
